Bug fix in height, width, and length: they're not ints, but floats. This caused errors.

+++ furni/furni.go
@@ -3,6 +3,6 @@ package furni
 type Furni struct {
 	Name   string
 	Height float64
-	Width  int
-	Length int
+	Width  float64
+	Length float64
 }
